refactor(ui): 拆分资源库边界并完善主题详情自适应
